+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.0.2] - 2018-6-26
+
 ### ✏️ Added
 
 - Firewall support, allowed options [here](packages/hume-app#appstartpath-opts---null).